ajout desactivation utilisateur

+++ app/Http/Controllers/LoginController.php
@@ -104,19 +104,23 @@ class LoginController extends Controller
 
     public function login(Request $request)
     {
+        $user = User::where('email', '=', $request->email)->get('statut');
+        if($user[0]->statut === 0){
+            return back()->withErrors(['message' => 'Le compte est désactiver contactez votre administrateur pour plus d’informations.'])->withInput();
+        }
         //essaye de connecter l'utilisateur avec les informations transmises
         if (Auth::guard()->attempt($request->only('email', 'password'))) {
-            if(auth()->guard()->check()){
-                //suppression des tokens si utilisateur est connecté
-                $request->user()->tokens()->delete();
-            }
-            $request->session()->regenerate();
-            //créer le token de connexion
-            $token = Auth()->user()->createToken('auth_token')->plainTextToken;
-            $role = Auth::user()->type;
-            //ajout un role à l'utilisateur
-            Auth()->user()->assignRole($role);
-            return redirect('/home')->with('token','Bearer '.$token);
+                if(auth()->guard()->check()){
+                    //suppression des tokens si utilisateur est connecté
+                    $request->user()->tokens()->delete();
+                }
+                $request->session()->regenerate();
+                //créer le token de connexion
+                $token = Auth()->user()->createToken('auth_token')->plainTextToken;
+                $role = Auth::user()->type;
+                //ajout un role à l'utilisateur
+                Auth()->user()->assignRole($role);
+                return redirect('/home')->with('token','Bearer '.$token);
         }
         return back()->withErrors(['message'=>'Données de connexion invalides.'])->withInput();
     }
